+++ integration_tests/empty.go
@@ -1,0 +1,3 @@
+// Empty file so that go install ./... doesn't complain with 'no buildable source files'
+
+package integration
