--- integration_tests/fake_api/empty.go
@@ -1,1 +1,0 @@
-package fakeapi
