--- integration_tests/real_api/empty.go
@@ -1,1 +1,0 @@
-package realapi
